call system-configured browser instead of firefox
On Debian/Ubuntu systems, the alternatives infrastructure allows the
user to define the preferred default browser and updates a standard
symlink to that selection. This symlink is called 'x-www-browser'. By
calling this instead of firefox directly, a user can change the value of
the default browser and the googlereader plugin will continue to work
properly.